Changed user model, renamed controllers to services
