+++ models/user.mjs
@@ -1,16 +1,16 @@
 import mongoose from 'mongoose';
 
 const UserSchema = new mongoose.Schema({
+  email: {
+    type: String,
+    required: true,
+  },
   google: {
     id: {
       type: String,
       required: false,
     },
     name: {
-      type: String,
-      required: true,
-    },
-    email: {
       type: String,
       required: true,
     },
