+++ passport/index.mjs
@@ -2,7 +2,7 @@ import GoogleTokenStrategy from 'passport-google-id-token';
 import users from '../models/user';
 import passport from 'passport';
 import { GOOGLE_ID } from '../config';
-import * as userController from '../controllers/user';
+import * as userService from '../services/user';
 
 export default app => {
   passport.serializeUser((collection, done) => {
@@ -11,7 +11,7 @@ export default app => {
 
   passport.deserializeUser(async (user, done) => {
     try {
-      const user = await userController.findById(user._id);
+      const user = await userService.findById(user._id);
       done(null, user);
     } catch (err) {
       done(err, null);
@@ -25,23 +25,23 @@ export default app => {
       },
       async (parsedToken, googleId, done) => {
         try {
-          const user = await userController.findByGoogleId(googleId);
+          const user = await userService.findByGoogleId(googleId);
 
           if (user) {
             return done(null, user);
           }
 
-          const newUser = await userController.createUser({
+          const newUser = await userService.createUser({
+            email: parsedToken.payload.email,
             google: {
               id: googleId,
               name: parsedToken.payload.name,
-              email: parsedToken.payload.email,
             },
           });
 
           done(null, newUser);
         } catch (err) {
-          throw err;
+          done(err, null);
         }
       },
     ),
